# src/main/resources/static/hackhub-openapi.yaml
openapi: 3.0.0
info:
  title: hackhub-openapi
  version: 1.0.0
paths:
...
  /api/hackathons/:hackathon:
    get:
      summary: Restituisce un hackathon specifico
      operationId: >-
        c_users_heron_onedrive_documenti_bruno_hackhub_api_-_2_hackathon_restituisce_un_hackathon_specifico_yml
      description: ''
      tags:
        - Hackathon
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  nome:
                    type: string
                  luogo:
                    type: string
                  premioInDenaro:
                    type: integer
                  currency:
                    type: string
                  dimensioneMaxTeam:
                    type: integer
                  regolamento:
                    type: string
                  stato:
                    type: string
                  organizzatoreId:
                    type: string
                  giudiceId:
                    type: string
              example:
                id: ''
                nome: ''
                luogo: ''
                premioInDenaro: 0
                currency: ''
                dimensioneMaxTeam: 0
                regolamento: ''
                stato: ISCRIZIONE
                organizzatoreId: ''
                giudiceId: ''
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: ''
      parameters:
        - name: hackathon
          in: path
          description: ''
          required: true
          schema:
            type: string
          example: e8f9b1c2-8f80-4ef1-95df-462f3b4f11f3
...
  /api/hackathons/:hackathonId/teams/:teamId/submission:
    put:
      summary: Aggiorna una submission esistente
      operationId: >-
        c_users_heron_onedrive_documenti_bruno_hackhub_api_-_2_submission_aggiorna_una_submission_esistente_yml
      description: ''
      tags:
        - Submission
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  descrizione:
                    type: string
                  allegato:
                    type: string
                  voto:
                    type: integer
                  giudizio:
                    type: string
              example:
                name: ''
                descrizione: ''
                allegato: ''
                voto: 0
                giudizio: ''
      parameters:
        - name: hackathonId
          in: path
          description: ''
          required: true
          schema:
            type: string
          example: e8f9b1c2-8f80-4ef1-95df-462f3b4f11f3
        - name: teamId
          in: path
          description: ''
          required: true
          schema:
            type: string
          example: team fortissimo
      requestBody:
        $ref: '#/components/requestBodies/aggiorna_una_submission_esistente'
    post:
      summary: Invia una nuova submission
      operationId: >-
        c_users_heron_onedrive_documenti_bruno_hackhub_api_-_2_submission_invia_una_nuova_submission_yml
      description: ''
      tags:
        - Submission
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  descrizione:
                    type: string
                  allegato:
                    type: string
                  voto:
                    type: integer
                  giudizio:
                    type: string
              example:
                name: ''
                descrizione: ''
                allegato: ''
                voto: 0
                giudizio: ''
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: ''
      parameters:
        - name: hackathonId
          in: path
          description: ''
          required: true
          schema:
            type: string
          example: e8f9b1c2-8f80-4ef1-95df-462f3b4f11f3
        - name: teamId
          in: path
          description: ''
          required: true
          schema:
            type: string
          example: team fortissimo
      requestBody:
        $ref: '#/components/requestBodies/invia_una_nuova_submission'
    get:
      summary: Restituisce la submission di un team per un hackathon
      operationId: >-
        c_users_heron_onedrive_documenti_bruno_hackhub_api_-_2_submission_restituisce_la_submission_di_un_team_per_un_hackathon_yml
      description: ''
      tags:
        - Submission
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  descrizione:
                    type: string
                  allegato:
                    type: string
                  voto:
                    type: integer
                  giudizio:
                    type: string
              example:
                name: ''
                descrizione: ''
                allegato: ''
                voto: 0
                giudizio: ''
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: ''
      parameters:
        - name: hackathonId
          in: path
          description: ''
          required: true
          schema:
            type: string
          example: e8f9b1c2-8f80-4ef1-95df-462f3b4f11f3
        - name: teamId
          in: path
          description: ''
          required: true
          schema:
            type: string
          example: team fortissimo
  /api/hackathons/:hackathonId/teams/:teamId/submission/evaluate:
    post:
      summary: Assegna una valutazione a una submission
      operationId: >-
        c_users_heron_onedrive_documenti_bruno_hackhub_api_-_2_submission_assegna_una_valutazione_a_una_submission_yml
      description: ''
      tags:
        - Submission
      responses:
        '204':
          description: No Content
      parameters:
        - name: hackathonId
          in: path
          description: ''
          required: true
          schema:
            type: string
          example: e8f9b1c2-8f80-4ef1-95df-462f3b4f11f3
        - name: teamId
          in: path
          description: ''
          required: true
          schema:
            type: string
          example: team fortissimo
      requestBody:
        $ref: '#/components/requestBodies/assegna_una_valutazione_a_una_submission'
  /api/hackathons/:hackathonId/submission/all:
    get:
      summary: Restituisce tutte le submission di un hackathon
      operationId: >-
        c_users_heron_onedrive_documenti_bruno_hackhub_api_-_2_submission_restituisce_tutte_le_submission_di_un_hackathon_yml
      description: ''
      tags:
        - Submission
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    name:
                      type: string
                    descrizione:
                      type: string
                    allegato:
                      type: string
                    voto:
                      type: integer
                    giudizio:
                      type: string
              example:
                - name: ''
                  descrizione: ''
                  allegato: ''
                  voto: 0
                  giudizio: ''
      parameters:
        - name: hackathonId
          in: path
          description: ''
          required: true
          schema:
            type: string
          example: e8f9b1c2-8f80-4ef1-95df-462f3b4f11f3
...
servers:
  - url: http://localhost:8080
    description: Server di sviluppo
components:
  schemas:
...
    aggiorna_una_submission_esistente:
      type: object
      properties:
        descrizione:
          type: string
        allegato:
          type: string
      example:
        descrizione: nuova desc
        allegato: C:\Users\heron\OneDrive\Desktop\IntelliJ\HackHub
    assegna_una_valutazione_a_una_submission:
      type: object
      properties:
        voto:
          type: integer
        giudizio:
          type: string
      example:
        voto: 4
        giudizio: decente
    invia_una_nuova_submission:
      type: object
      properties:
        name:
          type: string
        descrizione:
          type: string
        allegato:
          type: string
      example:
        name: progetto
        descrizione: progetto vincitore
        allegato: C:\\Users\\heron\\OneDrive\\Desktop\\IntelliJ\\HackHub.jar
...
  requestBodies:
...
    aggiorna_una_submission_esistente:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/aggiorna_una_submission_esistente'
      description: ''
      required: true
    assegna_una_valutazione_a_una_submission:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/assegna_una_valutazione_a_una_submission'
      description: ''
      required: true
    invia_una_nuova_submission:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/invia_una_nuova_submission'
      description: ''
      required: true
...
  securitySchemes: {}